refactor: update all imports to use auth module directly
Source files:
  - invite_service.py: app.core.cognito → app.auth.cognito_admin
  - routes/admin.py: app.core.cognito → app.auth.cognito_admin,
                     app.api.deps → app.auth.deps

Unit tests (imports + patch targets):
  - test_cognito.py: patches → app.auth.cognito_token.settings/get_jwks
  - test_cognito_email_lookup.py: patches → app.auth.cognito_admin._get_cognito_client
  - test_cognito_invite.py: patches → app.auth.cognito_admin.settings/_get_cognito_client
  - test_auth_deps.py: import + patches → app.auth.deps
  - test_role_deps.py: import → app.auth.deps

Functional tests (patch targets):
  - test_admin.py, test_auth.py, test_orgs.py, test_admin_org_endpoints.py:
    all verify_token/get_or_create_user patches → app.auth.deps

All 197 tests pass.
