+++ docs/pr1
@@ -1,0 +1,36 @@
+PR 1 — refactor: extract auth module from core/cognito and api/deps
+Why first: This is the highest-risk split (hot-path token verification). Getting it merged and stable early de-risks everything else.
+
+Commits:
+
+chore: create domain directory skeleton
+
+mkdir -p app/{auth,orgs,admin,suppliers,products,users,invites,bootstrap,health}
+Add __init__.py to each new package
+refactor: split cognito.py into cognito_token.py and cognito_admin.py
+
+Move get_jwks, get_signing_key, verify_token, get_email_from_access_token → auth/cognito_token.py
+Move _get_cognito_client, invite_cognito_user, delete_cognito_user → auth/cognito_admin.py
+Leave core/cognito.py as a re-export shim (imports from new locations, re-exports all names) so nothing breaks yet
+refactor: extract auth deps from api/deps.py
+
+Move bearer_scheme, get_current_user, require_role → auth/deps.py
+Leave get_current_org_membership, require_org_role in api/deps.py for now
+api/deps.py imports from auth.deps and re-exports for backward compat
+refactor: move auth route to auth/routes.py
+
+Move api/routes/auth.py → auth/routes.py
+Update main.py router import
+refactor: update all imports to use auth module directly
+
+Replace all from app.core.cognito import ... with from app.auth.cognito_token import ... / from app.auth.cognito_admin import ...
+Replace all from app.api.deps import get_current_user / require_role with from app.auth.deps import ...
+Remove re-export shims from core/cognito.py and api/deps.py
+Update all test patches (patch("app.core.cognito.verify_token") → patch("app.auth.cognito_token.verify_token"), etc.)
+chore: delete core/cognito.py and api/routes/auth.py
+
+Files touched: core/cognito.py, api/deps.py, api/routes/auth.py, main.py, api/routes/admin.py, api/routes/orgs.py, services/invite_service.py, ~15 test files that patch cognito/deps
+
+PR description:
+
+Splits core/cognito.py (monolith mixing hot-path JWT verification with cold-path Cognito admin calls) into auth/cognito_token.py (request-critical, cached JWKS + verify) and auth/cognito_admin.py (invite/delete — rare, network-heavy). Extracts auth-related FastAPI dependencies (get_current_user, require_role) from api/deps.py into auth/deps.py. Moves auth route into auth/routes.py. Uses intermediate re-export shims to ensure every commit is green.
