+++ docs/pr3
@@ -1,0 +1,44 @@
+PR 3 — refactor: migrate products, suppliers, invites, bootstrap, health
+Why third: These are straightforward file moves with minimal cross-cutting risk. Quick win after the two harder PRs.
+
+Commits:
+
+refactor: move products routes, schemas, and service to products/
+
+api/routes/products.py → products/routes.py
+schemas/product.py → products/schemas.py
+services/product_service.py → products/service.py
+Update imports in moved files + callers
+refactor: move suppliers routes, schemas, and service to suppliers/
+
+api/routes/suppliers.py → suppliers/routes.py
+schemas/supplier.py → suppliers/schemas.py
+services/supplier_service.py → suppliers/service.py
+refactor: move invite service to invites/
+
+services/invite_service.py → invites/service.py
+Update imports in callers (orgs/routes.py, admin/routes_orgs.py)
+refactor: extract bootstrap seeder and seed data
+
+core/bootstrap.py logic → bootstrap/seeder.py
+SEED_SUPPLIERS + SEED_PRODUCTS dicts → bootstrap/seed_data.py
+Update main.py lifespan import
+refactor: move health route to health/
+
+api/routes/health.py → health/routes.py
+Move schemas/user.py → auth/schemas.py (it only contains UserResponse, used by auth route)
+refactor: update main.py router registration for new structure
+
+Update all router imports in main.py to point to domain packages
+Final confirmation that all routes are correctly wired
+chore: remove empty legacy directories
+
+Delete api (all routes, deps, permissions moved)
+Delete schemas (all schemas moved)
+Delete services (all services moved)
+Delete core/bootstrap.py
+Update test patches for any remaining old paths
+PR description:
+
+Completes the domain-based restructure by migrating products, suppliers, invites, bootstrap, and health into their own packages. Extracts seed data from bootstrap logic into a dedicated seed_data.py. Removes the now-empty api, schemas, and services directories. After this PR, every domain is self-contained: {domain}/routes.py, {domain}/schemas.py, {domain}/service.py.
+
