+++ docs/pr4
@@ -1,0 +1,27 @@
+PR 4 — infra: set up real-database test infrastructure in testsv2/
+Why separate: Decouples the structural refactor from the test philosophy change. Existing tests remain untouched and passing.
+
+Commits:
+
+chore: create testsv2 directory with conftest and factories
+
+Create testsv2/conftest.py — real Postgres session fixtures with transaction-per-test rollback (per Section 4.2 of the plan)
+Create testsv2/factories.py — factory functions that INSERT real rows: create_user, create_org, create_membership, create_supplier, create_product, create_product_supplier, create_org_with_owner (per Section 4.3)
+Create testsv2/__init__.py
+Create testsv2/integration/__init__.py and testsv2/functional/__init__.py
+chore: add functional test auth bypass fixture
+
+Create testsv2/functional/conftest.py — bypass_auth fixture + AUTH_HEADER constant (per Section 4.5)
+chore: update Makefile and CI for testsv2
+
+Add test-db Makefile target (spin up local test Postgres on port 5433)
+Add test-v2 Makefile target: DATABASE_URL=... alembic upgrade head && pytest testsv2/ -v
+Update pyproject.toml — keep testpaths = ["tests"] so existing suite is unchanged; test-v2 target uses explicit path
+Update ci.yml — add alembic upgrade head step before tests, add a second pytest step for testsv2/
+test: add smoke integration test to validate fixture setup
+
+Create testsv2/integration/test_smoke.py — one or two tests that create a user/org via factories and query them back, proving the transaction rollback fixture works end-to-end
+PR description:
+
+Sets up testsv2/ with real-database test infrastructure using PostgreSQL and transaction-per-test rollback. Includes factory functions for all models, auth bypass fixtures for functional tests, and CI/Makefile integration. Existing tests in tests are untouched. Future PRs will incrementally add integration and functional tests here and eventually migrate coverage from tests.
+
