+++ docs/pr5
@@ -1,0 +1,28 @@
+PR 5 — docs: update README and add post-migration checklist
+Commits:
+
+docs: update README with new project structure
+
+Replace the project structure section with the new domain-based layout
+Document the dependency rules (Section 7 of the plan)
+Add the "Adding a New Feature" checklist (Section 8)
+docs: add testing guide and local test DB setup
+
+Document the testing philosophy: mock external services, never mock your own database
+Document make test-db for local Postgres setup
+Document make test-v2 for running real-DB tests
+Explain the testsv2/ migration strategy
+chore: archive refactoring plan as completed
+
+Move/rename EasyInventory_API_Refactoring_Plan (1).md.md) to indicate completion status
+PR description:
+
+Updates project documentation to reflect the new domain-based architecture. Adds developer guides for the testing strategy, local test database setup, and the "Adding a New Feature" checklist. Archives the refactoring plan document.
+
+PR Merge Order & Branch Strategy
+PR	Branch	Base	Can parallelize?
+PR 1	refactor/extract-auth-module	development	—
+PR 2	refactor/extract-orgs-admin-users	development (after PR 1)	No — depends on PR 1
+PR 3	refactor/migrate-remaining-domains	development (after PR 2)	No — depends on PR 2
+PR 4	infra/testsv2-real-db	development (after PR 3)	Could branch from PR 1 but import paths change in PR 2/3, so best after PR 3
+PR 5	docs/post-refactor-update	development (after PR 4)	Yes — could go after PR 3
