+++ tests/unit/test_auth_deps.py
@@ -3,7 +3,8 @@ from unittest.mock import patch, MagicMock, AsyncMock
 from fastapi import HTTPException
 from fastapi.security import HTTPAuthorizationCredentials
 
-from app.api.deps import get_current_org_membership, get_current_user
+from app.auth.deps import get_current_user
+from app.api.deps import get_current_org_membership
 from app.models.org_membership import OrgMembership
 from app.models.user import User
 
@@ -22,7 +23,7 @@ async def test_invalid_token_raises_401():
         scheme="Bearer",
         credentials="bad-token",
     )
-    with patch("app.api.deps.verify_token", side_effect=Exception("Invalid")):
+    with patch("app.auth.deps.verify_token", side_effect=Exception("Invalid")):
         with pytest.raises(HTTPException) as exc_info:
             await get_current_user(credentials=creds, db=mock_db)
     assert exc_info.value.status_code == 401
@@ -39,10 +40,10 @@ async def test_valid_token_returns_user():
         credentials="valid-token",
     )
     with patch(
-        "app.api.deps.verify_token",
+        "app.auth.deps.verify_token",
         return_value={"sub": "abc-123", "email": "test@test.com"},
     ):
-        with patch("app.api.deps.get_or_create_user", return_value=mock_user):
+        with patch("app.auth.deps.get_or_create_user", return_value=mock_user):
             user = await get_current_user(credentials=creds, db=mock_db)
     assert user.cognito_sub == "abc-123"
     assert user.email == "test@test.com"
@@ -55,7 +56,7 @@ async def test_expired_token_raises_401():
         credentials="expired-token",
     )
     with patch(
-        "app.api.deps.verify_token",
+        "app.auth.deps.verify_token",
         side_effect=Exception("Signature has expired"),
     ):
         with pytest.raises(HTTPException) as exc_info:
